refactor(presentation): extract active-run timeline builder and tool identity utilities
- Move active-run timeline projection logic from timeline.ts into new
  active-run-timeline.ts for clearer ownership of live tail rendering
- Add timeline-tool-identity.ts with tool call keying, consumption tracking,
  and live-status helpers used by the active-run builder
- Simplify ComposerToolbar permission options (remove iconClass, clarify
  icon usage is for tab trigger only)
- Update timeline.split-builders.test.ts to match refactored structure
