+++ packages/workbench-app/src/lib/presentation/state/timeline.split-builders.test.ts
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import { describe, it } from "node:test";
+import { buildActiveRunTimeline } from "./active-run-timeline";
 import {
-  buildActiveRunTimeline,
   buildCommittedTimeline,
   buildConversationTimeline,
   selectVisibleCommitted,
